requirements.txt
# Server dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
watchdog>=4.0.0
aiofiles>=23.2.1
mutagen>=1.47.0
httpx>=0.27.0
pydantic-settings>=2.0.0
yt-dlp>=2026.3.17

# Utility script dependencies (sunfly_match.py)
pdfplumber>=0.11.0
